Improved shift+rclick behavior when adding waypoints
Now when you shift+rclick a waypoint is added to the current path,
instead of entering a "waypoint mode" the resets the path.

There's still a bug: If you add a waypoint after the unit has passed
any waypoints, it will return to those waypoints. This might be annoying
but it's still better than how it used to be.

It's possible to perfect the shift+rclick behavior, but that would
require changing the protocol: "SEND_WAYPOINTS" should be made into
"CLEAR_WAYPOINTS" and "ADD_WAYPOINT". But that would break compatibility
with other clients, so I'm not in a rush to do that.